+++ src/Property/AtRule.php
@@ -13,14 +13,14 @@ interface AtRule extends Renderable, Commentable
      *
      * @var string
      */
-    const BLOCK_RULES = 'media/document/supports/region-style/font-feature-values';
+    public const BLOCK_RULES = 'media/document/supports/region-style/font-feature-values';
 
     /**
      * … and more font-specific ones (to be used inside font-feature-values)
      *
      * @var string
      */
-    const SET_RULES = 'font-face/counter-style/page/swash/styleset/annotation';
+    public const SET_RULES = 'font-face/counter-style/page/swash/styleset/annotation';
 
     /**
      * @return string|null