+++ src/Property/KeyframeSelector.php
@@ -9,7 +9,7 @@ class KeyframeSelector extends Selector
      *
      * @var string
      */
-    const SELECTOR_VALIDATION_RX = '/
+    public const SELECTOR_VALIDATION_RX = '/
     ^(
         (?:
             [a-zA-Z0-9\x{00A0}-\x{FFFF}_^$|*="\'~\[\]()\-\s\.:#+>]* # any sequence of valid unescaped characters