+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Changed
 
+- Add visibility to all class/interface constants (#469)
+
 ### Deprecated
 
 ### Removed